+++ src/Component/AbstractComponent.php
@@ -3,9 +3,9 @@
 namespace Solarium\Component;
 
 use Solarium\Component\RequestBuilder\ComponentRequestBuilderInterface;
-use Solarium\Component\ResponseParser\ComponentParserInterface;
 use Solarium\Core\Configurable;
 use Solarium\Core\Query\AbstractQuery;
+use Solarium\Core\Query\Helper;
 
 /**
  * Query component base class.
@@ -32,11 +32,11 @@ abstract class AbstractComponent extends Configurable
     abstract public function getRequestBuilder();
 
     /**
-     * Get the response parser class for this query.
-     *
-     * @return ComponentParserInterface
+     * This component has no response parser...
      */
-    abstract public function getResponseParser();
+    public function getResponseParser()
+    {
+    }
 
     /**
      * Set parent query instance.
@@ -60,5 +60,19 @@ abstract class AbstractComponent extends Configurable
     public function getQueryInstance()
     {
         return $this->queryInstance;
+    }
+
+    /**
+     * Returns a query helper.
+     *
+     * @return \Solarium\Core\Query\Helper
+     */
+    public function getHelper()
+    {
+        if ($queryInstance = $this->getQueryInstance()) {
+            return $this->getQueryInstance()->getHelper();
+        } else {
+            return new Helper();
+        }
     }
 }